+++ tests/Provider/ManagerTest.php
@@ -1,7 +1,9 @@
 <?php
+declare( strict_types=1 );
 
 namespace Rarst\Laps\Tests\Provider;
 
+use function Brain\Monkey\Functions\expect;
 use Pimple\Exception\FrozenServiceException;
 use Rarst\Laps\Plugin;
 use Rarst\Laps\Provider\Manager_Provider;
@@ -13,6 +15,8 @@ class ManagerTest extends LapsTestCase {
 	 * @covers \Rarst\Laps\Provider\Manager_Provider
 	 */
 	public function testProvider() {
+
+		expect( 'wp_normalize_path' )->zeroOrMoreTimes()->andReturnFirstArg();
 
 		$container = new Plugin();
 		$provider  = new Manager_Provider();